chore: Removed example C code and added a Rust library (ffi_plugin) to generate the necessary binaries.
During the creation process of a Flutter FFI plugin using the FFI Plugin template, the default example C code was removed. In its place, a Rust library called ffi_plugin was created, which will contain the necessary code to generate the binaries used in the plugin. This will allow for greater efficiency and flexibility in the plugin implementation.